+++ config/local.ts
@@ -6,7 +6,7 @@ const local: Config = {
     address: '0x7c2C195CD6D34B8F845992d380aADB2730bB9C6F'
   },
   token: {
-    totalSupply: '1000000000000000000',
+    totalSupply: '1000' + '0'.repeat(18),
     name: 'Umbrella',
     symbol: 'UMB'
   },