+++ contracts/interfaces/IStakingBank.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.6.8;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IStakingBank is IERC20 {
-  function receiveApproval(address _from, uint256 _value, bytes calldata _data) external returns (bool success);
+  function receiveApproval(address _from) external returns (bool success);
 
   function withdraw(uint256 _value) external returns (bool success);
 }