+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
-- option to remove validator
+- option to remove validator + update scripts
 
 ## [0.2.0] - 2021-03-10
 ### Added